Add stack clear command, fix midi bug
If you play the same note twice, the len auto noteoff would
sometimes cause the second note to get cut off
As a temp fix I set the noteoff to happen slightly early.
There can still be a situation however where you play again
way before the noteoff cb is triggered and the second note will get cutoff
when the cb hits. Not sure yet how (or if) I want to handle that.